+++ packages/components/CHANGELOG.md
@@ -3,6 +3,192 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.34](https://github.com/mittwald/flow/compare/0.1.0-alpha.33...0.1.0-alpha.34) (2024-03-14)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.33](https://github.com/mittwald/flow/compare/0.1.0-alpha.32...0.1.0-alpha.33) (2024-03-14)
+
+### Features
+
+- add color to Icon in Avatar
+  ([a430727](https://github.com/mittwald/flow/commit/a430727deb8f7290daf38fb0340d92401b2f4830))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.32](https://github.com/mittwald/flow/compare/0.1.0-alpha.31...0.1.0-alpha.32) (2024-03-13)
+
+### Bug Fixes
+
+- link component in link
+  ([959f6bb](https://github.com/mittwald/flow/commit/959f6bb554283010b2d271b245b4b017c4e7392f))
+
+### Features
+
+- add link component to link
+  ([b2ddea8](https://github.com/mittwald/flow/commit/b2ddea8a3bfc1596ba721fb92b6004e44823b54b))
+- add LinkProvider for Next.JS
+  ([1d275f4](https://github.com/mittwald/flow/commit/1d275f46be029179d4129b53b8c58fb0346e5fa5))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.31](https://github.com/mittwald/flow/compare/0.1.0-alpha.30...0.1.0-alpha.31) (2024-03-13)
+
+### Bug Fixes
+
+- fix navigation ([#222](https://github.com/mittwald/flow/issues/222))
+  ([bcbbe2e](https://github.com/mittwald/flow/commit/bcbbe2edd642989f6d195883d37bee153e3d0268))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.30](https://github.com/mittwald/flow/compare/0.1.0-alpha.29...0.1.0-alpha.30) (2024-03-12)
+
+### Features
+
+- allow text with icon in button
+  ([d20c356](https://github.com/mittwald/flow/commit/d20c3564e9e98befc3c8a5f28a241a74485a48d0))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.29](https://github.com/mittwald/flow/compare/0.1.0-alpha.28...0.1.0-alpha.29) (2024-03-12)
+
+### Features
+
+- add InlineCode component
+  ([c9fc931](https://github.com/mittwald/flow/commit/c9fc931777ae9d31f9f9c2f41f410af70cb13f40))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.28](https://github.com/mittwald/flow/compare/0.1.0-alpha.27...0.1.0-alpha.28) (2024-03-12)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.27](https://github.com/mittwald/flow/compare/0.1.0-alpha.26...0.1.0-alpha.27) (2024-03-12)
+
+### Features
+
+- add ColumnLayout
+  ([ac9a2c1](https://github.com/mittwald/flow/commit/ac9a2c1ac15bb80bfcddc614dc94198a8e4c7b5e))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.26](https://github.com/mittwald/flow/compare/0.1.0-alpha.24...0.1.0-alpha.26) (2024-03-12)
+
+### Features
+
+- add inline link
+  ([b237d0f](https://github.com/mittwald/flow/commit/b237d0f9a3f1ccb101366f0f5b43c3108c179439))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.25](https://github.com/mittwald/flow/compare/0.1.0-alpha.24...0.1.0-alpha.25) (2024-03-12)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.24](https://github.com/mittwald/flow/compare/0.1.0-alpha.23...0.1.0-alpha.24) (2024-03-12)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.23](https://github.com/mittwald/flow/compare/0.1.0-alpha.22...0.1.0-alpha.23) (2024-03-07)
+
+### Bug Fixes
+
+- update font sizes and line heights + rename story component categories
+  ([#210](https://github.com/mittwald/flow/issues/210))
+  ([abbc424](https://github.com/mittwald/flow/commit/abbc424077bb21e92a79ba77adebf0eb9c170d07))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.22](https://github.com/mittwald/flow/compare/0.1.0-alpha.21...0.1.0-alpha.22) (2024-03-07)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.21](https://github.com/mittwald/flow/compare/0.1.0-alpha.20...0.1.0-alpha.21) (2024-03-05)
+
+### Bug Fixes
+
+- token updates and small fixes
+  ([#208](https://github.com/mittwald/flow/issues/208))
+  ([85a8a1d](https://github.com/mittwald/flow/commit/85a8a1db4b4684552393d5fbdcf051defb39d85c))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.20](https://github.com/mittwald/flow/compare/0.1.0-alpha.19...0.1.0-alpha.20) (2024-03-01)
+
+### Features
+
+- add fonts
+  ([caf1bfc](https://github.com/mittwald/flow/commit/caf1bfc74c684f78b7656c244ff34aef7a3fd278))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.19](https://github.com/mittwald/flow/compare/0.1.0-alpha.18...0.1.0-alpha.19) (2024-02-29)
+
+### Features
+
+- new docs styling
+  ([90bd16d](https://github.com/mittwald/flow/commit/90bd16dd3d3021f993cb43c1e7c171340f9db09b))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
 # [0.1.0-alpha.18](https://github.com/mittwald/flow/compare/0.1.0-alpha.17...0.1.0-alpha.18) (2024-02-28)
 
 **Note:** Version bump only for package @mittwald/flow-react-components